Move tests directories
