+++ packages/protocol-kit/tests/e2e/contractManager.test.ts
@@ -1,8 +1,8 @@
 import chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
 import { deployments, waffle } from 'hardhat'
-import Safe, { ContractNetworksConfig } from '../src'
-import { ZERO_ADDRESS } from '../src/utils/constants'
+import Safe, { ContractNetworksConfig } from '../../src'
+import { ZERO_ADDRESS } from '../../src/utils/constants'
 import { getContractNetworks } from './utils/setupContractNetworks'
 import {
   getCompatibilityFallbackHandler,
