+++ packages/protocol-kit/tests/e2e/ethAdapters.test.ts
@@ -10,7 +10,7 @@ import {
   getSafeContractDeployment,
   getSafeProxyFactoryContractDeployment,
   getSignMessageLibContractDeployment
-} from '../src/contracts/safeDeploymentContracts'
+} from '../../src/contracts/safeDeploymentContracts'
 import { getContractNetworks } from './utils/setupContractNetworks'
 import {
   getCompatibilityFallbackHandler,
