+++ packages/protocol-kit/tests/e2e/execution.test.ts
@@ -8,8 +8,8 @@ import chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
 import { BigNumber } from 'ethers'
 import { deployments, waffle } from 'hardhat'
-import { safeVersionDeployed } from '../hardhat/deploy/deploy-contracts'
-import Safe from '../src'
+import { safeVersionDeployed } from '../../hardhat/deploy/deploy-contracts'
+import Safe from '../../src'
 import { itif } from './utils/helpers'
 import { getContractNetworks } from './utils/setupContractNetworks'
 import { getERC20Mintable, getSafeWithOwners } from './utils/setupContracts'
