+++ packages/protocol-kit/tests/e2e/guardManager.test.ts
@@ -1,39 +1,35 @@
 import chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
 import { deployments, waffle } from 'hardhat'
-import { safeVersionDeployed } from '../hardhat/deploy/deploy-contracts'
-import Safe, { SafeTransactionOptionalProps } from '../src'
-import { ZERO_ADDRESS } from './../src/utils/constants'
+import { safeVersionDeployed } from '../../hardhat/deploy/deploy-contracts'
+import Safe, { SafeTransactionOptionalProps } from '../../src'
+import { ZERO_ADDRESS } from '../../src/utils/constants'
 import { itif } from './utils/helpers'
 import { getContractNetworks } from './utils/setupContractNetworks'
-import {
-  getCompatibilityFallbackHandler,
-  getDefaultCallbackHandler,
-  getSafeWithOwners
-} from './utils/setupContracts'
+import { getDebugTransactionGuard, getSafeWithOwners } from './utils/setupContracts'
 import { getEthAdapter } from './utils/setupEthAdapter'
 import { getAccounts } from './utils/setupTestNetwork'
 import { waitSafeTxReceipt } from './utils/transactions'
 
 chai.use(chaiAsPromised)
 
-describe('Fallback handler manager', () => {
+describe('Safe guard manager', () => {
   const setupTests = deployments.createFixture(async ({ deployments }) => {
     await deployments.fixture()
     const accounts = await getAccounts()
     const chainId: number = (await waffle.provider.getNetwork()).chainId
     const contractNetworks = await getContractNetworks(chainId)
     return {
+      debugTransactionGuard: await getDebugTransactionGuard(),
       safe: await getSafeWithOwners([accounts[0].address]),
       accounts,
-      contractNetworks,
-      defaultCallbackHandler: await getDefaultCallbackHandler()
+      contractNetworks
     }
   })
 
-  describe('getFallbackHandler', async () => {
-    itif(safeVersionDeployed < '1.1.1')(
-      'should fail if getting the enabled fallback handler is not supported',
+  describe('getGuard', async () => {
+    itif(safeVersionDeployed < '1.3.0')(
+      'should fail if getting the enabled guard is not supported',
       async () => {
         const { safe, accounts, contractNetworks } = await setupTests()
         const [account1] = accounts
@@ -43,39 +39,19 @@ describe('Fallback handler manager', () => {
           safeAddress: safe.address,
           contractNetworks
         })
-        const tx = safeSdk.getFallbackHandler()
+        const tx = safeSdk.getGuard()
         await chai
           .expect(tx)
           .to.be.rejectedWith(
-            'Current version of the Safe does not support the fallback handler functionality'
+            'Current version of the Safe does not support Safe transaction guards functionality'
           )
       }
     )
 
-    itif(safeVersionDeployed >= '1.1.1')('should return the enabled fallback handler', async () => {
-      const { safe, accounts, contractNetworks, defaultCallbackHandler } = await setupTests()
-      const [account1] = accounts
-      const ethAdapter = await getEthAdapter(account1.signer)
-      const safeSdk = await Safe.create({
-        ethAdapter,
-        safeAddress: safe.address,
-        contractNetworks
-      })
-      const compatibilityFallbackHandler = (await getCompatibilityFallbackHandler()).contract
-        .address
-      chai.expect(await safeSdk.getFallbackHandler()).to.be.eq(compatibilityFallbackHandler)
-      const tx = await safeSdk.createEnableFallbackHandlerTx(defaultCallbackHandler.address)
-      const txResponse = await safeSdk.executeTransaction(tx)
-      await waitSafeTxReceipt(txResponse)
-      chai.expect(await safeSdk.getFallbackHandler()).to.be.eq(defaultCallbackHandler.address)
-    })
-  })
-
-  describe('createEnableFallbackHandlerTx', async () => {
-    itif(safeVersionDeployed < '1.1.1')(
-      'should fail if enabling a fallback handler is not supported',
+    itif(safeVersionDeployed >= '1.3.0')(
+      'should return 0x address when no Safe guard is enabled',
       async () => {
-        const { safe, accounts, contractNetworks, defaultCallbackHandler } = await setupTests()
+        const { safe, accounts, contractNetworks } = await setupTests()
         const [account1] = accounts
         const ethAdapter = await getEthAdapter(account1.signer)
         const safeSdk = await Safe.create({
@@ -83,16 +59,49 @@ describe('Fallback handler manager', () => {
           safeAddress: safe.address,
           contractNetworks
         })
-        const tx = safeSdk.createEnableFallbackHandlerTx(defaultCallbackHandler.address)
+        chai.expect(await safeSdk.getGuard()).to.be.eq(ZERO_ADDRESS)
+      }
+    )
+
+    itif(safeVersionDeployed >= '1.3.0')('should return the enabled Safe guard', async () => {
+      const { safe, accounts, contractNetworks, debugTransactionGuard } = await setupTests()
+      const [account1] = accounts
+      const ethAdapter = await getEthAdapter(account1.signer)
+      const safeSdk = await Safe.create({
+        ethAdapter,
+        safeAddress: safe.address,
+        contractNetworks
+      })
+      chai.expect(await safeSdk.getGuard()).to.be.eq(ZERO_ADDRESS)
+      const tx = await safeSdk.createEnableGuardTx(debugTransactionGuard.address)
+      const txResponse = await safeSdk.executeTransaction(tx)
+      await waitSafeTxReceipt(txResponse)
+      chai.expect(await safeSdk.getGuard()).to.be.eq(debugTransactionGuard.address)
+    })
+  })
+
+  describe('createEnableGuardTx', async () => {
+    itif(safeVersionDeployed < '1.3.0')(
+      'should fail if enabling a Safe guard is not supported',
+      async () => {
+        const { safe, accounts, contractNetworks, debugTransactionGuard } = await setupTests()
+        const [account1] = accounts
+        const ethAdapter = await getEthAdapter(account1.signer)
+        const safeSdk = await Safe.create({
+          ethAdapter,
+          safeAddress: safe.address,
+          contractNetworks
+        })
+        const tx = safeSdk.createEnableGuardTx(debugTransactionGuard.address)
         await chai
           .expect(tx)
           .to.be.rejectedWith(
-            'Current version of the Safe does not support the fallback handler functionality'
+            'Current version of the Safe does not support Safe transaction guards functionality'
           )
       }
     )
 
-    itif(safeVersionDeployed >= '1.1.1')('should fail if address is invalid', async () => {
+    itif(safeVersionDeployed >= '1.3.0')('should fail if address is invalid', async () => {
       const { safe, accounts, contractNetworks } = await setupTests()
       const [account1] = accounts
       const ethAdapter = await getEthAdapter(account1.signer)
@@ -101,11 +110,11 @@ describe('Fallback handler manager', () => {
         safeAddress: safe.address,
         contractNetworks
       })
-      const tx = safeSdk.createEnableFallbackHandlerTx('0x123')
-      await chai.expect(tx).to.be.rejectedWith('Invalid fallback handler address provided')
+      const tx = safeSdk.createEnableGuardTx('0x123')
+      await chai.expect(tx).to.be.rejectedWith('Invalid guard address provided')
     })
 
-    itif(safeVersionDeployed >= '1.1.1')(
+    itif(safeVersionDeployed >= '1.3.0')(
       'should fail if address is equal to 0x address',
       async () => {
         const { safe, accounts, contractNetworks } = await setupTests()
@@ -116,13 +125,13 @@ describe('Fallback handler manager', () => {
           safeAddress: safe.address,
           contractNetworks
         })
-        const tx = safeSdk.createEnableFallbackHandlerTx(ZERO_ADDRESS)
-        await chai.expect(tx).to.be.rejectedWith('Invalid fallback handler address provided')
+        const tx = safeSdk.createEnableGuardTx(ZERO_ADDRESS)
+        await chai.expect(tx).to.be.rejectedWith('Invalid guard address provided')
       }
     )
 
-    itif(safeVersionDeployed >= '1.1.1')('should fail if address is already enabled', async () => {
-      const { safe, accounts, contractNetworks, defaultCallbackHandler } = await setupTests()
+    itif(safeVersionDeployed >= '1.3.0')('should fail if address is already enabled', async () => {
+      const { safe, accounts, contractNetworks, debugTransactionGuard } = await setupTests()
       const [account1] = accounts
       const ethAdapter = await getEthAdapter(account1.signer)
       const safeSdk = await Safe.create({
@@ -130,17 +139,17 @@ describe('Fallback handler manager', () => {
         safeAddress: safe.address,
         contractNetworks
       })
-      const tx1 = await safeSdk.createEnableFallbackHandlerTx(defaultCallbackHandler.address)
+      const tx1 = await safeSdk.createEnableGuardTx(debugTransactionGuard.address)
       const txResponse = await safeSdk.executeTransaction(tx1)
       await waitSafeTxReceipt(txResponse)
-      const tx2 = safeSdk.createEnableFallbackHandlerTx(defaultCallbackHandler.address)
-      await chai.expect(tx2).to.be.rejectedWith('Fallback handler provided is already enabled')
+      const tx2 = safeSdk.createEnableGuardTx(debugTransactionGuard.address)
+      await chai.expect(tx2).to.be.rejectedWith('Guard provided is already enabled')
     })
 
-    itif(safeVersionDeployed >= '1.1.1')(
+    itif(safeVersionDeployed >= '1.3.0')(
       'should build the transaction with the optional props',
       async () => {
-        const { safe, accounts, contractNetworks, defaultCallbackHandler } = await setupTests()
+        const { safe, accounts, contractNetworks, debugTransactionGuard } = await setupTests()
         const [account1] = accounts
         const ethAdapter = await getEthAdapter(account1.signer)
         const safeSdk = await Safe.create({
@@ -156,10 +165,7 @@ describe('Fallback handler manager', () => {
           nonce: 555,
           safeTxGas: 666
         }
-        const tx = await safeSdk.createEnableFallbackHandlerTx(
-          defaultCallbackHandler.address,
-          options
-        )
+        const tx = await safeSdk.createEnableGuardTx(debugTransactionGuard.address, options)
         chai.expect(tx.data.baseGas).to.be.eq(111)
         chai.expect(tx.data.gasPrice).to.be.eq(222)
         chai.expect(tx.data.gasToken).to.be.eq('0x333')
@@ -169,8 +175,8 @@ describe('Fallback handler manager', () => {
       }
     )
 
-    itif(safeVersionDeployed >= '1.1.1')('should enable a fallback handler', async () => {
-      const { safe, accounts, contractNetworks, defaultCallbackHandler } = await setupTests()
+    itif(safeVersionDeployed >= '1.3.0')('should enable a Safe guard', async () => {
+      const { safe, accounts, contractNetworks, debugTransactionGuard } = await setupTests()
       const [account1] = accounts
       const ethAdapter = await getEthAdapter(account1.signer)
       const safeSdk = await Safe.create({
@@ -178,19 +184,17 @@ describe('Fallback handler manager', () => {
         safeAddress: safe.address,
         contractNetworks
       })
-      const compatibilityFallbackHandler = (await getCompatibilityFallbackHandler()).contract
-        .address
-      chai.expect(await safeSdk.getFallbackHandler()).to.be.eq(compatibilityFallbackHandler)
-      const tx = await safeSdk.createEnableFallbackHandlerTx(defaultCallbackHandler.address)
+      chai.expect(await safeSdk.getGuard()).to.be.eq(ZERO_ADDRESS)
+      const tx = await safeSdk.createEnableGuardTx(debugTransactionGuard.address)
       const txResponse = await safeSdk.executeTransaction(tx)
       await waitSafeTxReceipt(txResponse)
-      chai.expect(await safeSdk.getFallbackHandler()).to.be.eq(defaultCallbackHandler.address)
+      chai.expect(await safeSdk.getGuard()).to.be.eq(debugTransactionGuard.address)
     })
   })
 
-  describe('createDisableFallbackHandlerTx', async () => {
-    itif(safeVersionDeployed < '1.1.1')(
-      'should fail if disabling a fallback handler is not supported',
+  describe('createDisableGuardTx', async () => {
+    itif(safeVersionDeployed < '1.3.0')(
+      'should fail if disabling a Safe guard is not supported',
       async () => {
         const { accounts, contractNetworks } = await setupTests()
         const [account1] = accounts
@@ -201,41 +205,32 @@ describe('Fallback handler manager', () => {
           safeAddress: safe.address,
           contractNetworks
         })
-        const tx = safeSdk.createDisableFallbackHandlerTx()
+        const tx = safeSdk.createDisableGuardTx()
         await chai
           .expect(tx)
           .to.be.rejectedWith(
-            'Current version of the Safe does not support the fallback handler functionality'
+            'Current version of the Safe does not support Safe transaction guards functionality'
           )
       }
     )
 
-    itif(safeVersionDeployed >= '1.1.1')(
-      'should fail if no fallback handler is enabled',
-      async () => {
-        const { safe, accounts, contractNetworks } = await setupTests()
-        const [account1] = accounts
-        const ethAdapter = await getEthAdapter(account1.signer)
-        const safeSdk = await Safe.create({
-          ethAdapter,
-          safeAddress: safe.address,
-          contractNetworks
-        })
+    itif(safeVersionDeployed >= '1.3.0')('should fail if no Safe guard is enabled', async () => {
+      const { safe, accounts, contractNetworks } = await setupTests()
+      const [account1] = accounts
+      const ethAdapter = await getEthAdapter(account1.signer)
+      const safeSdk = await Safe.create({
+        ethAdapter,
+        safeAddress: safe.address,
+        contractNetworks
+      })
+      const tx = safeSdk.createDisableGuardTx()
+      await chai.expect(tx).to.be.rejectedWith('There is no guard enabled yet')
+    })
 
-        const tx = await safeSdk.createDisableFallbackHandlerTx()
-        const txResponse = await safeSdk.executeTransaction(tx)
-        await waitSafeTxReceipt(txResponse)
-        chai.expect(await safeSdk.getFallbackHandler()).to.be.eq(ZERO_ADDRESS)
-
-        const tx2 = safeSdk.createDisableFallbackHandlerTx()
-        await chai.expect(tx2).to.be.rejectedWith('There is no fallback handler enabled yet')
-      }
-    )
-
-    itif(safeVersionDeployed >= '1.1.1')(
+    itif(safeVersionDeployed >= '1.3.0')(
       'should build the transaction with the optional props',
       async () => {
-        const { accounts, contractNetworks, defaultCallbackHandler } = await setupTests()
+        const { accounts, contractNetworks, debugTransactionGuard } = await setupTests()
         const [account1] = accounts
         const safe = await getSafeWithOwners([account1.address])
         const ethAdapter = await getEthAdapter(account1.signer)
@@ -244,11 +239,10 @@ describe('Fallback handler manager', () => {
           safeAddress: safe.address,
           contractNetworks
         })
-
-        const tx1 = await safeSdk.createEnableFallbackHandlerTx(defaultCallbackHandler.address)
+        const tx1 = await safeSdk.createEnableGuardTx(debugTransactionGuard.address)
         const txResponse1 = await safeSdk.executeTransaction(tx1)
         await waitSafeTxReceipt(txResponse1)
-        chai.expect(await safeSdk.getFallbackHandler()).to.be.eq(defaultCallbackHandler.address)
+        chai.expect(await safeSdk.getGuard()).to.be.eq(debugTransactionGuard.address)
         const options: SafeTransactionOptionalProps = {
           baseGas: 111,
           gasPrice: 222,
@@ -257,7 +251,7 @@ describe('Fallback handler manager', () => {
           nonce: 555,
           safeTxGas: 666
         }
-        const tx2 = await safeSdk.createDisableFallbackHandlerTx(options)
+        const tx2 = await safeSdk.createDisableGuardTx(options)
         chai.expect(tx2.data.baseGas).to.be.eq(111)
         chai.expect(tx2.data.gasPrice).to.be.eq(222)
         chai.expect(tx2.data.gasToken).to.be.eq('0x333')
@@ -267,8 +261,8 @@ describe('Fallback handler manager', () => {
       }
     )
 
-    itif(safeVersionDeployed >= '1.1.1')('should disable an enabled fallback handler', async () => {
-      const { accounts, contractNetworks, defaultCallbackHandler } = await setupTests()
+    itif(safeVersionDeployed >= '1.3.0')('should disable an enabled Safe guard', async () => {
+      const { accounts, contractNetworks, debugTransactionGuard } = await setupTests()
       const [account1] = accounts
       const safe = await getSafeWithOwners([account1.address])
       const ethAdapter = await getEthAdapter(account1.signer)
@@ -278,15 +272,15 @@ describe('Fallback handler manager', () => {
         contractNetworks
       })
 
-      const tx = await safeSdk.createEnableFallbackHandlerTx(defaultCallbackHandler.address)
+      const tx = await safeSdk.createEnableGuardTx(debugTransactionGuard.address)
       const txResponse = await safeSdk.executeTransaction(tx)
       await waitSafeTxReceipt(txResponse)
-      chai.expect(await safeSdk.getFallbackHandler()).to.be.eq(defaultCallbackHandler.address)
+      chai.expect(await safeSdk.getGuard()).to.be.eq(debugTransactionGuard.address)
 
-      const tx1 = await safeSdk.createDisableFallbackHandlerTx()
+      const tx1 = await safeSdk.createDisableGuardTx()
       const txResponse1 = await safeSdk.executeTransaction(tx1)
       await waitSafeTxReceipt(txResponse1)
-      chai.expect(await safeSdk.getFallbackHandler()).to.be.eq(ZERO_ADDRESS)
+      chai.expect(await safeSdk.getGuard()).to.be.eq(ZERO_ADDRESS)
     })
   })
 })
