+++ packages/protocol-kit/tests/e2e/offChainSignatures.test.ts
@@ -5,8 +5,8 @@ import {
 import chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
 import { deployments, waffle } from 'hardhat'
-import { safeVersionDeployed } from '../hardhat/deploy/deploy-contracts'
-import Safe from '../src'
+import { safeVersionDeployed } from '../../hardhat/deploy/deploy-contracts'
+import Safe from '../../src'
 import { itif } from './utils/helpers'
 import { getContractNetworks } from './utils/setupContractNetworks'
 import { getSafeWithOwners } from './utils/setupContracts'
