+++ packages/protocol-kit/tests/e2e/ownerManager.test.ts
@@ -1,8 +1,8 @@
 import chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
 import { deployments, waffle } from 'hardhat'
-import Safe, { SafeTransactionOptionalProps } from '../src'
-import { SENTINEL_ADDRESS, ZERO_ADDRESS } from '../src/utils/constants'
+import Safe, { SafeTransactionOptionalProps } from '../../src'
+import { SENTINEL_ADDRESS, ZERO_ADDRESS } from '../../src/utils/constants'
 import { getContractNetworks } from './utils/setupContractNetworks'
 import { getSafeWithOwners } from './utils/setupContracts'
 import { getEthAdapter } from './utils/setupEthAdapter'
