+++ packages/protocol-kit/tests/e2e/safeFactory.test.ts
@@ -1,7 +1,7 @@
 import chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
 import { deployments, waffle } from 'hardhat'
-import { safeVersionDeployed } from '../hardhat/deploy/deploy-contracts'
+import { safeVersionDeployed } from '../../hardhat/deploy/deploy-contracts'
 import {
   ContractNetworksConfig,
   DeploySafeProps,
@@ -9,9 +9,9 @@ import {
   SafeAccountConfig,
   SafeDeploymentConfig,
   SafeFactory
-} from '../src'
-import { SAFE_LAST_VERSION } from '../src/contracts/config'
-import { ZERO_ADDRESS } from '../src/utils/constants'
+} from '../../src'
+import { SAFE_LAST_VERSION } from '../../src/contracts/config'
+import { ZERO_ADDRESS } from '../../src/utils/constants'
 import { itif } from './utils/helpers'
 import { getContractNetworks } from './utils/setupContractNetworks'
 import {
