+++ packages/protocol-kit/tests/e2e/transactionUtils.test.ts
@@ -1,7 +1,7 @@
 import { OperationType } from '@safe-global/safe-core-sdk-types'
 import chai from 'chai'
 
-import { decodeMultiSendData } from '../src/utils/transactions/utils'
+import { decodeMultiSendData } from '../../src/utils/transactions/utils'
 
 describe('Transaction utils', () => {
   describe('decodeMultiSend', () => {
