+++ packages/protocol-kit/tests/e2e/utilsSignatures.test.ts
@@ -1,5 +1,5 @@
 import chai from 'chai'
-import { adjustVInSignature, isTxHashSignedWithPrefix } from '../src/utils/signatures'
+import { adjustVInSignature, isTxHashSignedWithPrefix } from '../../src/utils/signatures'
 
 const safeTxHash = '0x4de27e660bd23052b71c854b0188ef1c5b325b10075c70f27afe2343e5c287f5'
 const signerAddress = '0xbc2BB26a6d821e69A38016f3858561a1D80d4182'
